[style] #108 컴포넌트 분리
